INTEGER -> BIGINT

+++ bot/sqlite_database/storage_sqlite.py
@@ -29,7 +29,7 @@ class StorageSqlite(Storage):
                     CREATE TABLE IF NOT EXISTS users
                     (
                         id INTEGER PRIMARY KEY,
-                        telegram_id INTEGER NOT NULL UNIQUE,
+                        telegram_id BIGINT NOT NULL UNIQUE,
                         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                         state TEXT DEFAULT NULL,
                         order_json TEXT DEFAULT NULL
